encoded length of a packet can be >255 bytes in case of FEC

+++ stack/framework/hal/chips/netdev_driver/inc/sx127x_netdev.h
@@ -35,7 +35,7 @@ extern const netdev_driver_t sx127x_driver;
  * @brief   Received LoRa packet status information
  */
 typedef struct netdev_radio_packet_info {
-    uint8_t rssi;           /**< RSSI of a received packet */
+    int16_t rssi;           /**< RSSI of a received packet */
     uint8_t lqi;            /**< LQI of a received packet */
     int8_t snr;             /**< S/N ratio */
     uint32_t time_on_air;   /**< Time on air of a received packet (ms) */
